huggingface/diffusers/image-to-video/stable_video_diffusion/requirements.txt
optimum-rbln
opencv-python-headless==4.10.0.84
